mission-bravo-engineer/backend/requirements.txt
fastapi
uvicorn
websockets
google-genai
python-dotenv
google-adk
pydantic
a2a-sdk